fixed voor sybren

+++ UDC/Source/UDC/Weapons/Weapon.h
@@ -34,7 +34,7 @@ public:
 	// Called every frame
 	virtual void Tick(float DeltaTime) override;
 
-	int val1;
-	int val2;
+	float val1;
+	float val2;
 
 };
